TST: cover blark CLI args
